(react): fixes starter-kit images error

+++ react/pro/starter-kit-generation/helpers.js
@@ -162,7 +162,9 @@ const copyRecursiveSync = (src, dest) => {
       )
     })
   } else {
-    fs.copyFileSync(src, dest)
+    if(fs.existsSync(src)){
+      fs.copyFileSync(src, dest)
+    }
   }
 }
